requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
bleach==6.0.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-cleanup==7.0.0
django-etc==1.4.0
django-hitcount==1.3.5
django-summernote==0.8.20.0
django-taggit==3.1.0
environs==9.5.0
idna==3.4
kombu==5.2.4
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.0
Pillow==9.3.0
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==5.3.0
PyMySQL==1.0.2